Quick fix for QUERY (find bases by army id, int -> long)

+++ src/main/java/com/vyshyvan/controllers/MilitaryBaseController.java
@@ -117,7 +117,7 @@ public class MilitaryBaseController {
     }
 
     @RequestMapping("/findAllByArmyId")
-    public List<MilitaryBase> findAllByArmyId(@RequestParam("id") int id){
+    public List<MilitaryBase> findAllByArmyId(@RequestParam("id") Long id){
         return militaryBaseService.findAllByArmyId(id);
     }
